Add files via upload
Sorting is an important data structure in many applications in the real-world. Several sorting algorithms are currently in use for searching and other operations. Sorting algorithms rearrange the elements of an array or list based on the elements’ comparison operators. The comparison operator is used in the accurate data structure to establish the new order of elements. This report analyzes and compares the time complexity and running time theoretically and experimentally of insertion, merge, and heap sort algorithms. Java language is used by NetBeans tool to implement the code of the algorithms. The results show that when dealing with sorted elements, insertion sort has a faster running time than merge and heap algorithms. When it comes to dealing with a large number of elements, it is better to use the merge sort. For the number of comparisons for each algorithm, the insertion sort has the highest number of comparisons.
by Khaznah Alhajri , Wala Alsinan , Sahar Almuhaishi , Fatimah Al-hmood , Narjis Al-Jumaia
Supervised by: Dr. Azza Ahmed Ali